se agrega interface para aplicar el principio de Open-Close.

+++ src/open_close/ShapesPrinter.java
@@ -2,7 +2,7 @@ package open_close;
 
 public class ShapesPrinter {
 
-    public String json(int sum) {
+    public String json(double sum) {
         return "{shapesSum: %s}".formatted(sum);
     }
